Lots of cleanup.
Most of it is such that the code still looks nice
when limited to an 80-character terminal.
I also used relative imports, which makes sense
with the reorganized repository.

Note that because of the relative imports,
tests have to run in "module" mode.
This can be done, for example, by
`cd`ing to the `nef` folder and executing

    python -m nef.test.test_array

to run the test_array.py file. Note the lack of .py
in the above command.