Added some informative documents and `setup.py`.
With `setup.py`, you can now install nef-py to a site-packages
directory using the command

    python setup.py install

To do this locally (which is likely preferable at the moment) do

    python setup.py install --user

Then, you can run scripts using nef-py by just importing
nef or nef.nef_theano as nef, and off you go.

This also means that we can upload to PyPI, but let's hold
off on that for a while.